+++ docs/changes.md
@@ -9,6 +9,7 @@ Support Change Log
 * Allow `Orchestra\Support\Nesty` to prepend an item without knowing the current first item.
 * Add `Orchestra\Support\Messages::extend()` and tweak how Messages notification can be manipulated on current request.
 * Add `Orchestra\Support\Nesty::is()` to return instance of `Illuminate\Support\Fluent` to allow further chaining of the instance.
+* Add `Orchestra\Support\Relic`.
 
 ## Version 2.0